Add context param to mount/route/render
An optional additional argument to mount, route and render that will
be passed through to a second argument when creating closure and
class components. This will enable writing more testable, loosly
coupled components.

This initial commit is for exploration of the idea and validating
performance impact. Here are my findings.

- This can be done as a non-breaking change.
- No noticable impact to memory usage or rendering performance.
  (Tested with DBMonster example for a render-heavy test case)
- Only adds 48 bytes when minified and gzipped.